Feat 1. DetailGoalColorFunc 파일 제거 및 만다라트뷰모델로 이동 2. 만부기 누르면 멘트에 맞게 말풍선 크기 변경되도록 구현
